Fix compiler warnings about converting size_t or mwIndex to int
svn-id: 2192 @trunk

+++ Ipopt/contrib/MatlabInterface/src/iterate.cpp
@@ -19,14 +19,14 @@ Iterate::Iterate (mxArray* ptr)
   if (mxIsCell(ptr)) {
 
     // The MATLAB array is a cell array. Repeat for each cell.
-    int n = mxGetNumberOfElements(ptr);
+    int n = (int) mxGetNumberOfElements(ptr);
     for (int i = 0; i < n; i++) {
       p = mxGetCell(ptr,i);  // Get the ith cell.
       if (!mxIsDouble(p) || mxIsComplex(p) || mxIsSparse(p))
 	throw MatlabException("The initial iterate must be either a full real array \
 in DOUBLE precision, or a cell array in which each cell is a full real array in \
 DOUBLE precision");      
-      nv += mxGetNumberOfElements(p);
+      nv += (int) mxGetNumberOfElements(p);
     }
   } else {
     
@@ -35,7 +35,7 @@ DOUBLE precision");
       throw MatlabException("The initial iterate must be either a full real array \
 in DOUBLE precision, or a cell array in which each cell is a full real array in \
 DOUBLE precision");
-    nv = mxGetNumberOfElements(ptr);
+    nv = (int) mxGetNumberOfElements(ptr);
   }
 }
 
@@ -45,10 +45,10 @@ void Iterate::inject (const double* x) {
     // The MATLAB array is a cell array. Repeat for each cell.
     mxArray* p; 
     int      m;
-    int      n = mxGetNumberOfElements(ptr);
+    int      n = (int) mxGetNumberOfElements(ptr);
     for (int i = 0; i < n; i++) {
       p = mxGetCell(ptr,i);  // Get the ith cell.
-      m = mxGetNumberOfElements(p);
+      m = (int) mxGetNumberOfElements(p);
       copymemory(x,mxGetPr(p),m);
       x += m;
     }
@@ -64,10 +64,10 @@ void Iterate::copyto (double* x) const {
     // The MATLAB array is a cell array. Repeat for each cell.
     const mxArray* p; 
     int            m;
-    int            n = mxGetNumberOfElements(ptr);
+    int            n = (int) mxGetNumberOfElements(ptr);
     for (int i = 0; i < n; i++) {
       p = mxGetCell(ptr,i);  // Get the ith cell.
-      m = mxGetNumberOfElements(p);
+      m = (int) mxGetNumberOfElements(p);
       copymemory(mxGetPr(p),x,m);
       x += m;
     }
